Sereal::Merger: suppress compilation warnings

+++ Perl/Merger/srl_merger.c
@@ -315,7 +315,7 @@ srl_merger_append(pTHX_ srl_merger_t *mrg, SV *src)
     /* preallocate space in obuf,
      * but this is still not enough because due to
      * varint we might need more space in obug then size of ibuf */
-    GROW_BUF(mrg->obuf, BUF_SIZE(mrg->ibuf));
+    GROW_BUF(mrg->obuf, (size_t) BUF_SIZE(mrg->ibuf));
 
     mrg->ibuf.pos = mrg->ibuf.body_pos + 1;
     srl_merge_single_value(mrg);
